+++ library/hls/src/main/java/com/google/android/exoplayer2/source/hls/playlist/HlsMediaPlaylist.java
@@ -139,7 +139,7 @@ public final class HlsMediaPlaylist extends HlsPlaylist {
    * The media sequence number of the first media segment in the playlist, as defined by
    * #EXT-X-MEDIA-SEQUENCE.
    */
-  public final int mediaSequence;
+  public final long mediaSequence;
   /**
    * The compatibility version, as defined by #EXT-X-VERSION.
    */
@@ -196,11 +196,23 @@ public final class HlsMediaPlaylist extends HlsPlaylist {
    * @param initializationSegment See {@link #initializationSegment}.
    * @param segments See {@link #segments}.
    */
-  public HlsMediaPlaylist(@PlaylistType int playlistType, String baseUri, List<String> tags,
-      long startOffsetUs, long startTimeUs, boolean hasDiscontinuitySequence,
-      int discontinuitySequence, int mediaSequence, int version, long targetDurationUs,
-      boolean hasIndependentSegmentsTag, boolean hasEndTag, boolean hasProgramDateTime,
-      DrmInitData drmInitData, Segment initializationSegment, List<Segment> segments) {
+  public HlsMediaPlaylist(
+      @PlaylistType int playlistType,
+      String baseUri,
+      List<String> tags,
+      long startOffsetUs,
+      long startTimeUs,
+      boolean hasDiscontinuitySequence,
+      int discontinuitySequence,
+      long mediaSequence,
+      int version,
+      long targetDurationUs,
+      boolean hasIndependentSegmentsTag,
+      boolean hasEndTag,
+      boolean hasProgramDateTime,
+      DrmInitData drmInitData,
+      Segment initializationSegment,
+      List<Segment> segments) {
     super(baseUri, tags);
     this.playlistType = playlistType;
     this.startTimeUs = startTimeUs;